split up tests and extend examples
