+++ examples/private-repository-with-team/versions.tf
@@ -1,0 +1,11 @@
+# ---------------------------------------------------------------------------------------------------------------------
+# SET TERRAFORM AND PROVIDER REQUIREMENTS FOR RUNNING THIS MODULE
+# ---------------------------------------------------------------------------------------------------------------------
+
+terraform {
+  required_version = ">= 0.12.9"
+
+  required_providers {
+    github = ">= 2.3.1, < 3.0.0"
+  }
+}
